+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Changed MSRV to 1.77.0.
 - Rate-limit portable actor registrations.
 - Raise data integrity error if webfinger hostname is missing in remote profile data.
+- Made unsolicited message check more efficient.
 
 ### Fixed